fix(refactor): fix getProjectByEnvironmentId helper and problems resolver references
